add logo lockup option
